+++ Sources/Swift/Core/Integrations/Integrations.swift
@@ -36,8 +36,12 @@ private struct AnyIntegration {
     @objc public class func install(with options: Options) {
         let dependencies = SentryDependencyContainer.sharedInstance()
 
+        // The order of integrations here is important.
+        // SentryCrashIntegration needs to be initialized before SentryAutoSessionTrackingIntegration.
+        // And SentrySessionReplayIntegration before SentryCrashIntegration.
         var integrations: [AnyIntegration] = [
             .init(SwiftAsyncIntegration.self),
+            .init(SentryCrashIntegration.self),
             .init(SentryAutoSessionTrackingIntegration.self),
             .init(SentryNetworkTrackingIntegration.self),
             .init(SentryHangTrackerIntegrationObjC.self),
@@ -59,6 +63,7 @@ private struct AnyIntegration {
 
         #if (os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)) && !SENTRY_NO_UIKIT
         integrations.append(.init(SentryScreenshotIntegration.self))
+        integrations.append(.init(SentryViewHierarchyIntegration.self))
         #endif
         
         #if os(iOS) || os(macOS)